Implement foreign keys management regarding graphs
The feature triggered an alternative
needed implementation for update methods.
Test informally for now through a REST client.
Foreign key for constraints is still missing.

+++ src/it/polito/verefoo/rest/spring/service/GraphService.java
@@ -72,7 +72,11 @@ public class GraphService {
         }
 
         public void deleteGraphs() {
-                graphRepository.deleteAll();
+                if (graphRepository.isAnyReferred()) {
+                        // throw an exception
+                } else {
+                        graphRepository.deleteAll();
+                }
         }
 
         /**
@@ -82,14 +86,58 @@ public class GraphService {
          * @param graph
          * @return a new id for the modified graph
          */
-        public Long updateGraph(Long gid, Graph graph) {
-                graphRepository.deleteById(gid);
-                DbGraph dbGraph = graphRepository.save(converter.deserializeGraph(graph));
-                return dbGraph.getId();
+        public void updateGraph(Long id, Graph graph) {
+                // possible alternative implementation through merge
+                // deserialize the graph into dbGraph
+                // set the ids of dbGraph as the ids of the existing graph
+                // store dbGraph
+
+                DbGraph newDbGraph = converter.deserializeGraph(graph);
+
+                DbGraph oldDbGraph;
+                Optional<DbGraph> dbGraph = graphRepository.findById(id);
+                if (dbGraph.isPresent())
+                        oldDbGraph = dbGraph.get();
+                else
+                        return;
+                
+                // merge
+                newDbGraph.setId(id);
+                graphRepository.save(newDbGraph, 0);
+                if (newDbGraph.getNode().size() >= oldDbGraph.getNode().size()) {
+                        int i = 0;
+                        for ( ; i < oldDbGraph.getNode().size(); i++) {
+                                // newDbGraph.getNode().get(i).setId(oldDbGraph.getNode().get(i).getId());
+                                updateNode(id, oldDbGraph.getNode().get(i).getId(), graph.getNode().get(i));
+                        }
+                        for ( ; i < newDbGraph.getNode().size(); i++) {
+                                createNode(id, graph.getNode().get(i));
+                        }
+                } else {
+                        int i = 0;
+                        for ( ; i < newDbGraph.getNode().size(); i++) {
+                                updateNode(id, oldDbGraph.getNode().get(i).getId(), graph.getNode().get(i));
+                        }
+                        for ( ; i < oldDbGraph.getNode().size(); i++) {
+                                // This solution doesn't work because the save method is inspired to MERGE,
+                                // so the nodes not referenced are not deleted
+                                // oldDbGraph.getNode().remove(i);
+                                deleteNode(oldDbGraph.getId(), oldDbGraph.getNode().get(i).getId()); 
+                        }
+                }
+
+                // old version: update by deleting and creating again, but a new id is generated
+                // graphRepository.deleteById(id);
+                // DbGraph dbGraph = graphRepository.save(converter.deserializeGraph(graph));
+                // return dbGraph.getId();
         }
 
-        public void deleteGraph(Long gid) {
-                graphRepository.deleteById(gid);
+        public void deleteGraph(Long id) {
+                if (graphRepository.isReferred(id)) {
+                        // throw exception (status code should be 409)
+                } else {
+                        graphRepository.deleteById(id);
+                }
         }
 
         public Graph getGraph(Long id) {
@@ -123,17 +171,51 @@ public class GraphService {
         /**
          * As usual, {@code node} and its neighbours must not declare any id.
          * 
-         * @param id
-         * @param nodeId
-         * @param node
-         * @return
          */
-        public Long updateNode(Long id, Long nodeId, Node node) {
-                deleteNode(id, nodeId);
-                return createNode(id, node);
+        @Transactional
+        public void updateNode(Long id, Long nodeId, Node node) {
+                DbNode newDbNode = converter.deserializeNode(node);
+                DbNode oldDbNode;
+                Optional<DbNode> dbNode = nodeRepository.findById(nodeId);
+                if (dbNode.isPresent()) {
+                        oldDbNode = dbNode.get();
+                } else
+                        return;
+                
+                // merge
+                newDbNode.setId(nodeId);
+                nodeRepository.save(newDbNode, 0);
+                updateConfiguration(id, nodeId, oldDbNode.getConfiguration().getId(), node.getConfiguration());
+                // neighbours can be updated by just deleting all of them and creating new ones, since
+                // the ids are not visible to the user
+                if (newDbNode.getNeighbour().size() >= oldDbNode.getNeighbour().size()) {
+                        int i = 0;
+                        for ( ; i < oldDbNode.getNeighbour().size(); i++) {
+                                // newDbGraph.getNode().get(i).setId(oldDbGraph.getNode().get(i).getId());
+                                updateNeighbour(id, nodeId, oldDbNode.getNeighbour().get(i).getId(), node.getNeighbour().get(i));
+                        }
+                        for ( ; i < newDbNode.getNeighbour().size(); i++) {
+                                createNeighbour(id, nodeId, node.getNeighbour().get(i));
+                        }
+                } else {
+                        int i = 0;
+                        for ( ; i < newDbNode.getNeighbour().size(); i++) {
+                                updateNeighbour(id, nodeId, oldDbNode.getNeighbour().get(i).getId(), node.getNeighbour().get(i));
+                        }
+                        for ( ; i < oldDbNode.getNeighbour().size(); i++) {
+                                // This solution doesn't work because the save method is inspired to MERGE,
+                                // so the nodes not referenced are not deleted
+                                // oldDbGraph.getNode().remove(i);
+                                deleteNeighbour(id, nodeId, oldDbNode.getNeighbour().get(i).getId()); 
+                        }
+                }
+
+                // old version: update by deleting and creating again, but a new id is generated
+                // deleteNode(id, nodeId);
+                // return createNode(id, node);
         }
 
-        public Node getNode(Long gid, Long nodeId) {
+        public Node getNode(Long id, Long nodeId) {
                 Optional<DbNode> dbNode = nodeRepository.findById(nodeId);
                 if (dbNode.isPresent()) {
                         return converter.serializeNode(dbNode.get());
@@ -147,7 +229,14 @@ public class GraphService {
                 return dbNeighbour.getId();
         }
 
-        public void deleteNeighbour(Long gid, Long nodeId, Long neighbourId) {
+        // TODO: decide whether to add a controller for this function or not
+        private void updateNeighbour(Long id, Long nodeId, Long neighbourId, Neighbour neighbour) {
+                DbNeighbour newDbNeighbour = converter.deserializeNeighbour(neighbour);
+                newDbNeighbour.setId(neighbourId);
+                neighbourRepository.save(newDbNeighbour);
+        }
+
+        public void deleteNeighbour(Long id, Long nodeId, Long neighbourId) {
                 nodeRepository.unbindNeighbour(nodeId, neighbourId);
                 neighbourRepository.deleteById(neighbourId);
         }
@@ -157,13 +246,20 @@ public class GraphService {
         }
 
         @Transactional
-        public Long updateConfiguration(Long id, Long nodeId, Long configurationId, Configuration configuration) {
-                nodeRepository.unbindConfiguration(nodeId);
-                configurationRepository.deleteById(configurationId);
-                DbConfiguration dbConfiguration = configurationRepository
-                                .save(converter.deserializeConfiguration(configuration));
-                nodeRepository.bindConfiguration(nodeId, dbConfiguration.getId());
-                return dbConfiguration.getId();
+        public void updateConfiguration(Long id, Long nodeId, Long configurationId, Configuration configuration) {
+                DbConfiguration newDbConfiguration = converter.deserializeConfiguration(configuration);
+
+                configurationRepository.deleteFunctionsById(configurationId);
+                newDbConfiguration.setId(configurationId);
+                configurationRepository.save(newDbConfiguration);
+
+                // old version: update by deleting and creating again, but a new id is generated
+                // nodeRepository.unbindConfiguration(nodeId);
+                // configurationRepository.deleteById(configurationId);
+                // DbConfiguration dbConfiguration = configurationRepository
+                //                 .save(converter.deserializeConfiguration(configuration));
+                // nodeRepository.bindConfiguration(nodeId, dbConfiguration.getId());
+                // return dbConfiguration.getId();
         }
 
         public void createConstraints(Long id, Constraints constraints) {
